javaDoc jobOfferService

+++ src/main/java/com/accenture/recrume/recruMe/model/Applicant.java
@@ -14,7 +14,7 @@ public class Applicant {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
 
     private String firstName;
     private String lastName;
